Removed second friction clamping
Removed a line that clamps the applied voltage of a motor to be 0 if it is below the friction of moving. This is incorrect as even if a motor's voltage is below the friction of the ground, it is still using some current. Directly clamped using the friction calcualted in the motor configs instead.